Added docker files and also fixed group_id issue in migration

+++ database/migrations/2025_11_04_155139_add_group_id_to_accounts_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('accounts', function (Blueprint $table) {
-            $table->foreignId('group_id')->nullable()->after('user_id')->constrained('account_groups')->onDelete('set null');
+            if (!Schema::hasColumn('accounts', 'group_id')) {
+                $table->foreignId('group_id')->nullable()->after('user_id')->constrained('account_groups')->onDelete('set null');
+            }
         });
     }
